error checking if DSI cannot be started in this directory

+++ dsi/dsi.py
@@ -42,6 +42,13 @@ class DSI():
         self.schema_tables = set()
         self.loaded_tables = set()
 
+        if "/" in filename:
+            create_bool = self.t.can_create_file_here(filename.rsplit("/", 1)[0])
+        else:
+            create_bool = self.t.can_create_file_here()
+        if create_bool is False:
+            raise RuntimeError("Cannot initialize DSI due to write permissions in this directory. Please try elsewhere.")
+
         if filename == ".temp_dsi.db" and os.path.exists(filename):
             os.remove(filename)
